SystemUI: Turn on the display briefly when resuming AOD
On some devices, transition from AOD_PAUSED state (when display is
turned off when the proximity sensor is covered) back into AOD state
(when it is no longer covered), is broken and results in screen staying
black until it is turned on using the power button. This can be worked
around by switching the display into the ON state briefly (1 sec)
before coming back into the DOZE_AOD state.
Note that this does not enter the fully on state (lockscreen), the only
visible change is in brightness.

To enable this workaround, enable this in the device's SystemUI overlay:
config_displayTurnsOnWhenResumingAod

Change-Id: I05ebd5402c0789dd095ea5b0e1af6204d84c5a65